requirements.txt
streamlit==0.90.0
roboflow-client==1.0.2
pandas==1.3.3
